Update NDK version AGAIN.
rlottie uses pixman code, which needs to compile with -no-integrated-as. 23.1.7779620 has a bug where -no-integrated-as does not work properly. This was fixed in build 8486889. However we cant update to r24 yet because -no-integrated-as is removed completely there. So we are stuck here until pixman is updated (or we do some work to just remove that dependency).